ecom-pipeline/requirements.txt
apache-airflow==2.8.1
psycopg2-binary==2.9.9
faker==22.2.0
pandas==2.2.0
